[137] Create EditFreakPage and add react-router to App.js
