+++ src/components/FreakTile/FreakTile.jsx
@@ -1,13 +1,14 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import './FreakTile.css';
+import { Link } from 'react-router-dom';
 
 function FreakTile({ id, picture, name }) {
   return (
-    <a href={ `/freaks/${ id }` } className="tile" data-testid="freak-tile">
+    <Link to={ `/freaks/${ id }/edit` } className="tile" data-testid="freak-tile">
       <img src={ picture } alt="Face of the Freak" className="tile__img" data-testid="freak-picture" />
       <h1 className="tile__name" data-testid="freak-name">{ name }</h1>
-    </a>
+    </Link>
   );
 }
