+++ src/components/FreaksGrid/FreaksGrid.scss
@@ -3,7 +3,7 @@
   grid-template-columns: repeat(2, 1fr);
   grid-gap: 1.87rem 1.87rem;
   margin: 0 auto;
-  padding: 1.87rem 0.625rem;
+  padding: 3.87rem 0.625rem;
   max-width: 30.62rem;
 }
 
@@ -21,4 +21,3 @@
     max-width: 64rem;
   }
 }
-
