+++ src/pages/EditFreakPage/EditFreakPage.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function EditFreakPage() {
+  return (
+    <div>
+      <p>Create a new Freak</p>
+    </div>
+  );
+}
+
+export default EditFreakPage;
